Reorganize Netlify function for deployment
