+++ netlify/functions/chatbot/requirements.txt
